# Convert AES-128-CTR to pre-parsed C++ circuit