- Add support for loading and generating pre-parsed C++ code for BristolFashion circuits (mostly copied from BristolFormat)
- Use a pre-parsed AES-128 circuit for AES_128_CTR_Calculator in order to improve initialization time, especially in environments where sscanf is relatively slow (45 -> 35 ms in a program that uses the AES_128_CTR_Calculator)