@wangxiao1254 2022-07-25
Thanks for your contribution. However, I'm not sure if it is a good idea to put aes_128.cpp as the source since it is essentially duplicated information with aes_128.txt. 

There is a pre-parsed AES from https://github.com/emp-toolkit/emp-tool/blob/master/CMakeLists.txt#L17 already. What's the difference?