@kdrag0n 2022-07-26
> Thanks for your contribution. However, I'm not sure if it is a good idea to put aes_128.cpp as the source since it is essentially duplicated information with aes_128.txt. 

I agree that the duplication isn't ideal, but I'm not sure what the alternative is. The txt.cpp file can be generated at compile-time with shell commands, but generating this is a bit more complicated and requires calling `BristolFashion::to_file` from C++.

> 
> There is a pre-parsed AES from https://github.com/emp-toolkit/emp-tool/blob/master/CMakeLists.txt#L17 already. What's the difference?

The txt.cpp + fmemopen approach is certainly faster than reading from a file, but in my testing, it's still considerably slower than this because the textual data is parsed with sscanf (character-by-character) every time AES_128_CTR_Calculator is initialized.

This pre-parsed circuit already has all the text parsed into numbers that are compiled into the library as constant data (similar to the output of `BristolFormat::to_file`) and offloads the cost of parsing to a one-time job, so the program never incurs the cost of sscanf. The cost can be significant in some slower environments.

Benchmarks of a program that uses AES_128_CTR_Calculator, before and after this change:

Host: 45 -> 35 ms
Slower environment: 350 -> 250 ms (mostly relevant here)