@wangxiao1254 2022-07-26
The `aes_128.txt.cpp` does not need fmemopen right? the array is compiled to a binary statically.