@kdrag0n 2022-07-27
The txt.cpp file looks like this:

```
#include "emp-tool/circuits/aes_128_ctr.h"
unsigned char emp_tool_circuits_files_bristol_fashion_aes_128_txt[] = {
  0x33, 0x36, 0x36, 0x36, 0x33, 0x20, 0x33, 0x36, 0x39, 0x31, 0x39, 0x0a,
  0x32, 0x20, 0x31, 0x32, 0x38, 0x20, 0x31, 0x32, 0x38, 0x20, 0x0a, 0x31,
  0x20, 0x31, 0x32, 0x38, 0x20, 0x0a, 0x0a, 0x32, 0x20, 0x31, 0x20, 0x31,
  0x32, 0x38, 0x20, 0x30, 0x20, 0x33, 0x33, 0x32, 0x35, 0x34, 0x20, 0x58,
  0x4f, 0x52, 0x0a, 0x32, 0x20, 0x31, 0x20, 0x31, 0x32, 0x39, 0x20, 0x31,
  0x20, 0x33, 0x33, 0x32, 0x35, 0x35, 0x20, 0x58, 0x4f, 0x52, 0x0a, 0x32,
  0x20, 0x31, 0x20, 0x31, 0x33, 0x30, 0x20, 0x32, 0x20, 0x33, 0x33, 0x32,
  0x35, 0x36, 0x20, 0x58, 0x4f, 0x52, 0x0a, 0x32, 0x20, 0x31, 0x20, 0x31,
```

It's compiled into the binary, but the embedded data is the content of aes_128.txt — still in textual form. fmemopen is used to open a virtual FILE pointing to this data so the text can be parsed with fscanf: https://github.com/emp-toolkit/emp-tool/blob/f1f46ffa9c2bdccc6ca1648dbaf4a2bb5c4465b7/emp-tool/circuits/aes_128_ctr.h#L108

https://github.com/emp-toolkit/emp-tool/blob/1345a1462b12fd41fa2f02bd603b6e8067caf53d/emp-tool/circuits/circuit_file.h#L166