@wangxiao1254 2022-08-03
I see you are right! I think the best way is to modify the existing code so that the generated array in the current code base does not need parsing anymore, right?